requirements.txt
pygame-ce
argon2-cffi